Renamespace sign in redirect response factory
