+++ src/Security/TokenAuthentication/AuthenticationEntryPoint.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Security\TokenAuthentication;
 
 use App\RedirectRoute\Factory as RedirectRouteFactory;
-use App\SignInRedirectResponse\Factory as SignInRedirectResponseFactory;
+use App\Response\Factory as SignInRedirectResponseFactory;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
